08-11-2023
08-11-2023

+++ lib/screens/user/home/regular_listing/model/get_horse_detail_model.dart
@@ -18,7 +18,7 @@ class HorseDetailsResponse {
   Opinion? opinion;
   Data? data;
   int? sales;
-  int? ratting;
+  double? ratting;
 
   HorseDetailsResponse({
     this.success,
